Rename model -> logic
