+++ src/main/java/com/codecool/quest/logic/Drawable.java
@@ -1,4 +1,4 @@
-package com.codecool.quest.model;
+package com.codecool.quest.logic;
 
 public interface Drawable {
     String getTileName();
