+++ src/main/java/com/codecool/quest/logic/Player.java
@@ -1,4 +1,4 @@
-package com.codecool.quest.model;
+package com.codecool.quest.logic;
 
 public class Player extends Actor {
     public Player(Cell cell) {
