+++ src/test/java/com/codecool/quest/logic/CellTest.java
@@ -1,4 +1,4 @@
-package com.codecool.quest.model;
+package com.codecool.quest.logic;
 
 import org.junit.jupiter.api.Test;
